pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpuniverse"
version = "1.0.0"
authors = [
    {name = "Salesforce Research"}
]
description = "A framework for developing and benchmarking AI agents using Model Control Protocol"
readme = "README.md"
license = {text = "3-Clause BSD"}
keywords = [
    "AI",
    "Agents", 
    "MCP",
    "benchmarking",
    "LLM",
    "machine-learning"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10,<4"
dependencies = [
    "requests==2.32.4",
    "pydantic==2.10.6",
    "pydantic[email]==2.10.6", 
    "schema==0.7.7",
    "mcp==1.4.1",
    "httpx==0.28.1",
    "click==8.1.8",
    "jinja2==3.1.6",
    "python-dotenv==1.0.1",
    "anyio==4.9.0",
    "openai==1.68.2",
    "anthropic==0.49.0",
    "mistralai==1.6.0",
    "pyyaml==6.0.2",
    "google-genai==1.16.1",
    "redis==6.1.0",
    "psycopg[binary]==3.2.9",
    "sqlalchemy[asyncio]==2.0.41",
    "fastapi==0.115.12",
    "bcrypt==4.3.0",
    "pyseto==1.8.4",
    "celery==5.5.3",
    "xai-sdk==1.0.0",
    "claude-code-sdk==0.0.20",
    "wikipedia-api==0.8.1",
    "mcp_server_fetch",
    "google-auth==2.38.0",
    "google-auth-oauthlib==1.2.1",
    "google-api-python-client",
    "mcp_server_calculator==0.1.1",
    "yfinance==0.2.61",
    "blender-mcp==1.1.3",
    "playwright==1.52.0",
    "mathutils==3.3.0",
    "uvicorn[standard]==0.34.0",
    "pytz==2024.2"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-postgresql",
    "pylint",
    "pre-commit",
]
web = [
    "uvicorn[standard]==0.34.0",
    "fastapi==0.115.12",
    "celery==5.5.3",
    "redis==6.1.0",
    "psycopg[binary]==3.2.9",
    "sqlalchemy[asyncio]==2.0.41",
]
dashboard = [
    "gradio>=5.42.0",
]

[project.urls]
Homepage = "https://github.com/SalesforceAIResearch/MCP-Universe"
Repository = "https://github.com/SalesforceAIResearch/MCP-Universe"

[tool.setuptools.package-dir]
mcpuniverse = "mcpuniverse"

[tool.setuptools.packages.find]
include = ["mcpuniverse*"]

[tool.setuptools.package-data]
mcpuniverse = [
    "agent/configs/*",
    "app/db/migration/*", 
    "benchmark/configs/*",
    "dashboard/static/*",
    "dashboard/templates/*",
    "mcp/configs/*"
]